hateos

+++ src/main/java/br/com/fiap/mypet/repository/AnimalRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 
 import br.com.fiap.mypet.models.Animal;
-public interface AnimalRepository extends JpaRepository<Animal, Long> {
+public interface AnimalRepository extends JpaRepository<Animal, Integer> {
 
 }
