+++ src/main/java/br/com/fiap/mypet/repository/PagamentoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.com.fiap.mypet.models.Pagamento;
 
 
-public interface PagamentoRepository extends JpaRepository<Pagamento,Long> {
+public interface PagamentoRepository extends JpaRepository<Pagamento,Integer> {
     
 }
