+++ src/main/java/br/com/fiap/mypet/repository/ServicoRepository.java
@@ -2,10 +2,7 @@ package br.com.fiap.mypet.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
 import br.com.fiap.mypet.models.Servico;
-
-
-public interface ServicoRepository extends JpaRepository<Servico, Long> {
+public interface ServicoRepository extends JpaRepository<Servico, Integer> {
     
 }
